generalize summary function; move test; finish filling in
