+++ src/neuroconv/__init__.py
@@ -2,4 +2,5 @@ from .basedatainterface import BaseDataInterface
 from .baseextractorinterface import BaseExtractorInterface
 from .basetemporalalignmentinterface import BaseTemporalAlignmentInterface
 from .nwbconverter import ConverterPipe, NWBConverter
+from .tools import get_data_interfaces_summaries
 from .tools.yaml_conversion_specification import run_conversion_from_yaml
