+++ src/neuroconv/datainterfaces/ecephys/blackrock/blackrockdatainterface.py
@@ -12,7 +12,7 @@ class BlackrockRecordingInterface(BaseRecordingExtractorInterface):
     :py:class:`~spikeinterface.extractors.BlackrockRecordingExtractor`."""
 
     display_name = "Blackrock Recording"
-    associated_suffixes = (".ns{stream_number}",)
+    associated_suffixes = (".ns0", ".ns1", ".ns2", ".ns3", ".ns4", ".ns5")
     info = "Interface for Blackrock recording data."
 
     @classmethod
@@ -67,6 +67,10 @@ class BlackrockRecordingInterface(BaseRecordingExtractorInterface):
 
 class BlackrockSortingInterface(BaseSortingExtractorInterface):
     """Primary data interface class for converting Blackrock spiking data."""
+
+    display_name = "Blackrock Sorting"
+    associated_suffixes = (".nev",)
+    info = "Interface for Blackrock sorting data."
 
     @classmethod
     def get_source_schema(cls) -> dict:
