+++ src/neuroconv/datainterfaces/ecephys/neuralynx/neuralynxdatainterface.py
@@ -14,6 +14,7 @@ class NeuralynxRecordingInterface(BaseRecordingExtractorInterface):
     :py:class:`~spikeinterface.extractors.NeuralynxRecordingExtractor`."""
 
     display_name = "Neuralynx Recording"
+    associated_suffixes = (".nse", ".ntt", ".nse", ".nev")
     info = "Interface for Neuralynx recording data."
 
     @classmethod
@@ -85,6 +86,10 @@ class NeuralynxRecordingInterface(BaseRecordingExtractorInterface):
 
 
 class NeuralynxSortingInterface(BaseSortingExtractorInterface):
+    display_name = "Neuralynx Sorting"
+    associated_suffixes = (".nse", ".ntt", ".nse", ".nev")
+    info = "Interface for Neuralynx sorting data."
+
     def __init__(self, folder_path: FolderPathType, sampling_frequency: float = None, verbose: bool = True):
         """_summary_
