+++ src/neuroconv/datainterfaces/ophys/basesegmentationextractorinterface.py
@@ -14,6 +14,8 @@ from ...utils import fill_defaults, get_base_schema, get_schema_from_hdmf_class
 class BaseSegmentationExtractorInterface(BaseExtractorInterface):
     """Parent class for all SegmentationExtractorInterfaces."""
 
+    keywords = ("segmentation", "roi", "cells")
+
     ExtractorModuleName = "roiextractors"
 
     def __init__(self, **source_data):
