+++ src/neuroconv/datainterfaces/ophys/tiff/tiffdatainterface.py
@@ -5,8 +5,9 @@ from ....utils import FilePathType
 class TiffImagingInterface(BaseImagingExtractorInterface):
     """Interface for multi-page TIFF files."""
 
-    help = "Interface for multi-page TIFF files."
     display_name = "TIFF Imaging"
+    associated_suffixes = (".tif", ".tiff")
+    info = "Interface for multi-page TIFF files."
 
     @classmethod
     def get_source_schema(cls) -> dict:
