+++ src/neuroconv/tools/__init__.py
@@ -1,6 +1,11 @@
 """Collection of all helper functions that require at least one external dependency (some being optional as well)."""
 
-from .importing import get_package
+from .importing import (
+    get_data_interfaces_summaries,
+    get_package,
+    get_package_version,
+    is_package_installed,
+)
 from .nwb_helpers import get_module
 from .path_expansion import LocalPathExpander
 from .processes import deploy_process
